refactor : 컴포넌트 테스트로 이동
